rename config related classes
