+++ nvflare/app_common/job/fed_app_config.py
@@ -16,12 +16,12 @@ from typing import List
 from nvflare.apis.executor import Executor
 from nvflare.apis.impl.controller import Controller
 from nvflare.apis.impl.wf_comm_server import WFCommServer
-from nvflare.app_common.job.base_app import BaseApp
+from nvflare.app_common.job.base_app_config import BaseAppConfig
 from nvflare.private.fed.client.client_json_config import _ExecutorDef
 from nvflare.private.fed.server.server_json_config import WorkFlow
 
 
-class ClientApp(BaseApp):
+class ClientAppConfig(BaseAppConfig):
     def __init__(self) -> None:
         super().__init__()
 
@@ -37,7 +37,7 @@ class ClientApp(BaseApp):
         self.executors.append(e)
 
 
-class ServerApp(BaseApp):
+class ServerAppConfig(BaseAppConfig):
     def __init__(self) -> None:
         super().__init__()
 
@@ -61,13 +61,13 @@ class ServerApp(BaseApp):
 
 
 class FedApp:
-    def __init__(self, server_app: ServerApp = None, client_app: ClientApp = None) -> None:
+    def __init__(self, server_app: ServerAppConfig = None, client_app: ClientAppConfig = None) -> None:
         super().__init__()
 
-        if server_app and not isinstance(server_app, ServerApp):
-            raise ValueError(f"server_app must be type of ServerApp, but got {server_app.__class__}")
-        if client_app and not isinstance(client_app, ClientApp):
-            raise ValueError(f"client_app must be type of ClientApp, but got {client_app.__class__}")
+        if server_app and not isinstance(server_app, ServerAppConfig):
+            raise ValueError(f"server_app must be type of ServerAppConfig, but got {server_app.__class__}")
+        if client_app and not isinstance(client_app, ClientAppConfig):
+            raise ValueError(f"client_app must be type of ClientAppConfig, but got {client_app.__class__}")
 
-        self.server_app: ServerApp = server_app
-        self.client_app: ClientApp = client_app
+        self.server_app: ServerAppConfig = server_app
+        self.client_app: ClientAppConfig = client_app
